+++ packages/vsce/CHANGELOG.md
@@ -4,9 +4,13 @@ All notable changes to the "zowe-native-proto-vsce" extension will be documented
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## Recent Changes
+
+- Added experimental native client for improved performance which can be enabled via a VS Code setting. [#833](https://github.com/zowe/zowe-native-proto/pull/833)
+
 ## `0.3.0`
 
-- Used new SDK groupings that align with zowex. [#807](https://github.com/zowe/zowe-native-proto/pull/808)
+- Used new SDK groupings that align with zowex. [#807](https://github.com/zowe/zowe-native-proto/issues/808)
 
 ## `0.2.4`